# internal/adapter/http/http.go
// Package http provides an HTTP adapter for the Keystone agent.
// It exposes a REST API for controlling the agent and its components.
package http

import (
	"context"
	"crypto/subtle"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"log"
	"net"
	"net/http"
	"os"
	"strconv"
	"strings"
	"time"

	"github.com/carlosprados/keystone/internal/adapter"
)

// maxRequestBody bounds the size of a request body the API will read into
// memory (recipes and plans are small TOML/JSON documents). It defends against
// memory-exhaustion DoS via an oversized POST. Override with
// KEYSTONE_MAX_REQUEST_BYTES.
const defaultMaxRequestBody int64 = 4 << 20 // 4 MiB

// Adapter implements the HTTP REST API for the agent.
type Adapter struct {
	addr    string
	token   string
	handler adapter.CommandHandler
	server  *http.Server
}

// Config holds the configuration for the HTTP adapter.
type Config struct {
	Addr  string // Listen address (e.g., "127.0.0.1:8080")
	Token string // Bearer token required for the API; empty = no auth (loopback only)
}

// New creates a new HTTP adapter.
func New(cfg Config, handler adapter.CommandHandler) *Adapter {
	return &Adapter{
		addr:    cfg.Addr,
		token:   cfg.Token,
		handler: handler,
	}
}

// isLoopbackAddr reports whether the listen address binds only the loopback
// interface. An empty host (e.g. ":8080") binds all interfaces and is therefore
// NOT loopback.
func isLoopbackAddr(addr string) bool {
	host, _, err := net.SplitHostPort(addr)
	if err != nil {
		host = addr
	}
	if host == "" {
		return false
	}
	if host == "localhost" {
		return true
	}
	ip := net.ParseIP(host)
	return ip != nil && ip.IsLoopback()
}

// withAuth wraps a handler to require a bearer token when one is configured.
// /healthz is always exempt so liveness probes work without credentials. The
// token comparison is constant-time to avoid leaking it through timing.
func (a *Adapter) withAuth(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if a.token == "" || r.URL.Path == "/healthz" {
			next.ServeHTTP(w, r)
			return
		}
		const prefix = "Bearer "
		got := strings.TrimPrefix(r.Header.Get("Authorization"), prefix)
		if subtle.ConstantTimeCompare([]byte(got), []byte(a.token)) != 1 {
			w.Header().Set("WWW-Authenticate", "Bearer")
			http.Error(w, "unauthorized", http.StatusUnauthorized)
			return
		}
		next.ServeHTTP(w, r)
	})
}

// Name returns the adapter identifier.
func (a *Adapter) Name() string {
	return "http"
}

func maxRequestBody() int64 {
	if v := os.Getenv("KEYSTONE_MAX_REQUEST_BYTES"); v != "" {
		if n, err := strconv.ParseInt(v, 10, 64); err == nil && n > 0 {
			return n
		}
	}
	return defaultMaxRequestBody
}

// readLimitedBody reads the request body capped at maxRequestBody. It replaces
// r.Body with a MaxBytesReader so an oversized upload is rejected instead of
// being buffered fully into memory. On overflow or read error it writes a 413
// and returns ok=false; callers must return immediately.
func readLimitedBody(w http.ResponseWriter, r *http.Request) ([]byte, bool) {
	r.Body = http.MaxBytesReader(w, r.Body, maxRequestBody())
	body, err := io.ReadAll(r.Body)
	if err != nil {
		http.Error(w, "request body too large or unreadable", http.StatusRequestEntityTooLarge)
		return nil, false
	}
	return body, true
}

// Start initializes the HTTP server and begins listening.
func (a *Adapter) Start(ctx context.Context) error {
	// Fail closed: never expose an unauthenticated API beyond loopback. The API
	// can apply plans and run lifecycle hooks (i.e. arbitrary code), so binding
	// it to a reachable interface without a token is a remote-code-execution
	// hole.
	if a.token == "" && !isLoopbackAddr(a.addr) {
		return fmt.Errorf("refusing to start HTTP API on non-loopback address %q without authentication: set KEYSTONE_API_TOKEN or bind to 127.0.0.1", a.addr)
	}

	mux := a.buildRouter()

	a.server = &http.Server{
		Addr:    a.addr,
		Handler: a.withAuth(mux),
		// Bound the request-read phase to defeat slowloris-style attacks that
		// hold connections open by trickling headers/body. WriteTimeout is
		// deliberately left unset: handlePlanApply runs the apply (downloads +
		// install hooks) synchronously and may legitimately take minutes before
		// the response is written; a write deadline would abort it mid-install.
		ReadHeaderTimeout: 10 * time.Second,
		ReadTimeout:       60 * time.Second,
		IdleTimeout:       120 * time.Second,
		MaxHeaderBytes:    1 << 20, // 1 MiB
	}

	log.Printf("[http] starting HTTP adapter on %s", a.addr)

	// Start server in goroutine
	errCh := make(chan error, 1)
	go func() {
		if err := a.server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			errCh <- err
		}
		close(errCh)
	}()

	// Give the server a moment to start or fail
	select {
	case err := <-errCh:
		return fmt.Errorf("http server failed to start: %w", err)
	case <-time.After(100 * time.Millisecond):
		// Server started successfully
		return nil
	}
}

// Stop gracefully shuts down the HTTP server.
func (a *Adapter) Stop(ctx context.Context) error {
	if a.server == nil {
		return nil
	}
	log.Printf("[http] stopping HTTP adapter")
	return a.server.Shutdown(ctx)
}

// handleRoot serves the landing page.
func (a *Adapter) handleRoot(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
	_, _ = w.Write([]byte("Keystone agent is running. See /healthz, /metrics and /v1/components\n"))
}

// handleHealthz returns the agent health status.
func (a *Adapter) handleHealthz(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(a.handler.GetHealth())
}

// handleComponents returns the list of managed components.
func (a *Adapter) handleComponents(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(a.handler.GetComponents())
}

// handleRecipes handles GET (list) and POST (add) for recipes.
func (a *Adapter) handleRecipes(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		list, err := a.handler.ListRecipes()
		if err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(list)

	case http.MethodPost:
		force := r.URL.Query().Get("force") == "true"
		content, ok := readLimitedBody(w, r)
		if !ok {
			return
		}
		name, version, err := a.handler.AddRecipe(string(content), force)
		if err != nil {
			http.Error(w, err.Error(), http.StatusConflict)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusCreated)
		_ = json.NewEncoder(w).Encode(map[string]any{
			"status":  "created",
			"name":    name,
			"version": version,
		})

	default:
		w.WriteHeader(http.StatusMethodNotAllowed)
	}
}

// handleRecipeDelete handles DELETE /v1/recipes/{name}/{version}
func (a *Adapter) handleRecipeDelete(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodDelete {
		w.WriteHeader(http.StatusMethodNotAllowed)
		return
	}

	path := strings.TrimPrefix(r.URL.Path, "/v1/recipes/")
	parts := strings.Split(path, "/")
	if len(parts) != 2 {
		http.Error(w, "invalid path, expected /v1/recipes/{name}/{version}", http.StatusBadRequest)
		return
	}

	name, version := parts[0], parts[1]
	if err := a.handler.DeleteRecipe(name, version); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

// handleComponentAction handles POST /v1/components/{name}:stop and :restart
func (a *Adapter) handleComponentAction(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		w.WriteHeader(http.StatusMethodNotAllowed)
		return
	}

	path := strings.TrimPrefix(r.URL.Path, "/v1/components/")
	var action string
	switch {
	case strings.HasSuffix(path, ":stop"):
		action = "stop"
	case strings.HasSuffix(path, ":restart"):
		action = "restart"
	default:
		w.WriteHeader(http.StatusNotFound)
		return
	}

	name := strings.TrimSuffix(path, ":"+action)
	name = strings.Trim(name, "/")
	if name == "" {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	switch action {
	case "stop":
		if err := a.handler.StopComponent(name); err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		w.WriteHeader(http.StatusNoContent)

	case "restart":
		// Dry-run: return planned order without executing
		if r.URL.Query().Get("dry") == "true" {
			result := a.handler.RestartComponentDry(name)
			w.Header().Set("Content-Type", "application/json")
			_ = json.NewEncoder(w).Encode(result)
			return
		}

		// Parse wait mode and timeout
		wait := strings.ToLower(r.URL.Query().Get("wait"))
		if wait == "" {
			wait = "pid"
		}
		timeout := parseDurationDefault(r.URL.Query().Get("timeout"), 60*time.Second)

		log.Printf("[http] component=%s wait=%s timeout=%v msg=restarting component", name, wait, timeout)

		result, err := a.handler.RestartComponent(name, wait, timeout)
		if err != nil {
			log.Printf("[http] component=%s error=%v msg=restart failed", name, err)
			if strings.Contains(err.Error(), "timed out") {
				http.Error(w, err.Error(), http.StatusGatewayTimeout)
			} else {
				http.Error(w, err.Error(), http.StatusInternalServerError)
			}
			return
		}

		log.Printf("[http] component=%s pid=%d msg=restart completed", name, result.PID)
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(result)
	}
}

// handlePlanStatus returns the current plan status.
func (a *Adapter) handlePlanStatus(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(a.handler.GetPlanStatus())
}

// handlePlanGraph returns the dependency graph.
func (a *Adapter) handlePlanGraph(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(a.handler.GetPlanGraph())
}

// handlePlanApply applies a deployment plan.
func (a *Adapter) handlePlanApply(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		w.WriteHeader(http.StatusMethodNotAllowed)
		return
	}

	body, ok := readLimitedBody(w, r)
	if !ok {
		return
	}

	dry := r.URL.Query().Get("dry") == "true"

	// Reject the legacy {"planPath": ...} form: letting a remote caller choose
	// an arbitrary filesystem path to load and execute is a local-file-inclusion
	// vector. Plans must be supplied as content over the wire.
	var legacy struct {
		PlanPath string `json:"planPath"`
	}
	if json.Unmarshal(body, &legacy) == nil && legacy.PlanPath != "" {
		http.Error(w, "planPath is no longer accepted; upload plan content instead", http.StatusBadRequest)
		return
	}

	if err := a.handler.ApplyPlanContent(string(body), dry); err != nil {
		// A plan the caller got wrong is a 4xx. The distinction is not cosmetic:
		// a 500 tells automation to retry, and a file with a typo in it will
		// still have that typo on the next attempt.
		status := http.StatusInternalServerError
		if errors.Is(err, adapter.ErrInvalidInput) {
			status = http.StatusBadRequest
		}
		http.Error(w, err.Error(), status)
		return
	}
	w.WriteHeader(http.StatusAccepted)
}

// handlePlanStop stops all running components.
func (a *Adapter) handlePlanStop(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		w.WriteHeader(http.StatusMethodNotAllowed)
		return
	}

	if err := a.handler.StopPlan(); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

// handlePlanReconcile repairs the plan in effect.
//
// A pass that decided to change nothing is a success with skipped=true, not an
// error: "an apply is already running" is an ordinary answer, and reporting it
// as a failure would train callers to ignore real ones.
func (a *Adapter) handlePlanReconcile(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		w.WriteHeader(http.StatusMethodNotAllowed)
		return
	}

	res, err := a.handler.ReconcileNow()
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(res)
}

func parseDurationDefault(s string, d time.Duration) time.Duration {
	if s == "" {
		return d
	}
	if dd, err := time.ParseDuration(s); err == nil && dd > 0 {
		return dd
	}
	return d
}